Enh #7 - refac profile header and edit
